% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xkcd.R
\name{print_xkcd}
\alias{print_xkcd}
\title{Print XKCD}
\usage{
print_xkcd(comic)
}
\arguments{
\item{comic}{\code{integer}. Comic number.}
}
\value{
the xkcd comic in your default Viewer
}
\description{
Print an XKCD comic in your default Viewer
}
\examples{
\dontrun{print_xkcd(comic = 12)}
}
