% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comic.R
\name{get_comic}
\alias{get_comic}
\title{Get XKCD Comic Data}
\usage{
get_comic(comic = NULL)
}
\arguments{
\item{comic}{\code{integer}. Comic number.}
}
\value{
a tibble with comic data
}
\description{
Download data from individual XKCD comics from
<https://xkcd.com/>.
}
\examples{
\dontrun{get_comic(comic = 614)}
}
